Parallelize create_pretraining_data.py
Friday 31/01/2020, I launched create_pretraining_data.py in a for loop
with os.system. Process crashed because I ran out of memory on the disk.
So I moved to /raid/antoloui on fullfat where now I have 3.8TB of space.
I took the opportunity to parallelize the script: now one CPU for each
of the 14 files to process.